extract taskcard component
